Criação da página de login.
Foi inserido os níveis de usuários do sistema e também foi implementada a parte de proteção das páginas.

+++ clientes.php
@@ -1,9 +1,22 @@
 <?php
 
-// Coletando dados da base de dados.
 
+// Importando arquivos.
 include("lib/conexao.php");
 
+// Conferindo se existe alguma sessão iniciada.
+if(!isset($_SESSION)) {
+    // Criando nova sessão
+    session_start();
+}
+
+// Conferindo se existe um cliente logado.
+if (!$_SESSION['usuario']) {
+    header(("Location: index.php"));
+    die();
+}
+
+// Coletando dados da base de dados.
 $sql_code = "SELECT * FROM clientes ORDER BY id ASC ";
 $query_clientes = $mysqli->query($sql_code) or die($mysqli->error);
 $num_clientes = $query_clientes->num_rows;
@@ -20,23 +33,27 @@ $num_clientes = $query_clientes->num_rows;
 <body>
     
     <h1>Lista de Clientes</h1>
-    <p><a href="cadastrar_clientes.php">Cadastrar novo cliente</a></p>
-
+    <?php if($_SESSION['admin']) { ?>
+        <p><a href="cadastrar_clientes.php">Cadastrar novo cliente</a></p>
+    <?php } ?>
     <table border="1" cellpadding = "10">
         <thead>
             <th>ID</th>
+            <th>Admin</th>
             <th>Foto</th>
             <th>Nome</th>
             <th>E-mail</th>
             <th>Telefone</th>
             <th>Data de nascimento</th>
             <th>Data de cadastro</th>
-            <th>Ações</th>
+            <?php if($_SESSION['admin']) { ?>
+                <th>Ações</th>
+            <?php } ?>
         </thead>
         <tbody>
              <?php if($num_clientes == 0 )  {?> <!-- Tratando caso que não haja clientes na base de dados. -->
                 <tr>
-                    <td colspan="7">Nenhum cliente foi cadastrado!</td>
+                    <td colspan="<?php if($_SESSION['admin']) echo 9; else echo 8; ?>">Nenhum cliente foi cadastrado!</td>
                 </tr>
 
                 <?php 
@@ -72,19 +89,24 @@ $num_clientes = $query_clientes->num_rows;
                     <!-- Exibindo os dados dos clientes -->
 
                     <td style="text-align: center;"><?php echo $cliente['id'] ?></td>
-                    <td><img height="50" src="<?php echo $cliente['foto']; ?>"></td>
+                    <td style="text-align: center;"><?php if($cliente['admin']) echo "Sim"; else echo "Não"; ?></td>
+                    <td><img width="60" height="55" src="<?php if(empty($cliente['foto'])) echo "arquivos/usuario.jpg"; else echo $cliente['foto'] ?>"></td>
                     <td><?php echo $cliente['nome'] ?></td>
                     <td><?php echo $cliente['email'] ?></td>
                     <td><?php echo $telefone ?></td>
                     <td style="text-align: center;" ><?php echo $Dt_nascimento ?></td>
                     <td><?php echo $Dt_cadastro ?></td>
-                    <td><a href="editar_clientes.php?id=<?php echo $cliente['id'] ?>">Editar</a> <a href="excluir_clientes.php?id=<?php echo $cliente['id'] ?>">Deletar</a></td> <!-- Criando links das páginas de manipulação do cliente. -->
-                
+                    <?php if($_SESSION['admin']) { ?>
+                        <td><a href="editar_clientes.php?id=<?php echo $cliente['id'] ?>">Editar</a> <a href="excluir_clientes.php?id=<?php echo $cliente['id'] ?>">Deletar</a></td> <!-- Criando links das páginas de manipulação do cliente. -->
+                    <?php } ?>
                 </tr>
                 <?php
                 }
             }?>
         </tbody>
     </table>
+    <p>
+        <a href="logout.php">Sair do sistema</a>
+    </p>
 </body>
 </html>
